Pretty massive refactor to standardize src/lib outputs
