--- src/lib/google/googleClient.js
@@ -1,7 +1,0 @@
-const { google } = require("googleapis");
-
-module.exports = new google.auth.OAuth2(
-  process.env.SHEETS_CLIENT_ID,
-  process.env.SHEETS_CLIENT_SECRET,
-  process.env.SHEETS_REDIRECT_URI
-);
