+++ src/scripts/setup.js
@@ -10,7 +10,6 @@ const {
   accountsSetupCompleted,
   sheetsSetupCompleted
 } = require('../lib/common')
-const { logPromise } = require('../lib/logging')
 const _ = require('lodash')
 
 try {
@@ -26,62 +25,80 @@ try {
     server.use(bodyParser.urlencoded({ extended: false }))
     server.use(bodyParser.json())
 
-    server.get('/config', async (req, res) => {
-      const config = await logPromise(getConfigEnv(), 'Serving current config')
-      return res.json({
-        ...config,
-        accountsSetupCompleted: accountsSetupCompleted(),
-        sheetsSetupCompleted: sheetsSetupCompleted()
-      })
-    })
-
-    server.put('/config', async (req, res) => {
-      return updateConfig({ [req.body.id]: req.body.value })
-    })
-
-    server.delete('/config', async (req, res) => {
-      return deleteConfigProperty(req.body.id)
-    })
-
-    server.get('/balances', async (req, res, next) => {
-      switch (process.env.ACCOUNT_PROVIDER) {
-        case 'plaid':
-          return res.json(await require('../lib/plaid/plaid').fetchBalances())
-        default:
-          return {}
-      }
-    })
-
-    server.post('/token', async (req, res, next) => {
-      switch (process.env.ACCOUNT_PROVIDER) {
-        case 'plaid':
-          return await require('../lib/plaid/plaid').saveAccessToken(req.body.public_token, req.body.accountNickname)
-        default:
-          return
-      }
-    })
-
-    server.post('/update', async (req, res, next) => {
-      switch (process.env.ACCOUNT_PROVIDER) {
-        case 'plaid':
-          return res.json({
-            public_token: await require('../lib/plaid/plaid').createPublicToken(
-              process.env[`PLAID_TOKEN_${nickname}`],
-              req.body.accountNickname
-            )
+    server.get('/config', (req, res) => {
+      return getConfigEnv()
+        .then(config =>
+          res.json({
+            data: {
+              ...config,
+              accountsSetupCompleted: accountsSetupCompleted(),
+              sheetsSetupCompleted: sheetsSetupCompleted()
+            }
           })
+        )
+        .catch(error => res.json(error))
+    })
+
+    server.put('/config', (req, res) => {
+      return updateConfig(req.body.updates)
+        .then(response => res.json({ data: {} }))
+        .catch(error => res.json(error))
+    })
+
+    server.delete('/config', (req, res) => {
+      return deleteConfigProperty(req.body.id)
+        .then(response => res.json({ data: {} }))
+        .catch(error => res.json(error))
+    })
+
+    server.get('/balances', (req, res) => {
+      switch (process.env.ACCOUNT_PROVIDER) {
+        case 'plaid':
+          return require('../lib/plaid')
+            .fetchBalances()
+            .then(balances => res.json({ data: balances }))
+            .catch(error => res.json(error))
         default:
-          return
+          return res.json({ data: {} })
       }
     })
 
-    server.get('/google-sheets-url', async (req, res) => {
-      return res.json({ url: require('../lib/google/sheets').getAuthURL() })
+    server.post('/token', (req, res) => {
+      switch (process.env.ACCOUNT_PROVIDER) {
+        case 'plaid':
+          return require('../lib/plaid')
+            .saveAccessToken(req.body.public_token, req.body.accountNickname)
+            .then(balances => res.json({ data: balances }))
+            .catch(error => res.json(error))
+        default:
+          return res.json({ data: {} })
+      }
     })
 
-    server.get('/google-sheets-oauth2callback', async (req, res) => {
-      await require('../lib/google/sheets').getToken(req.query.code)
-      return res.redirect('http://localhost:3000/sheets')
+    server.post('/update', (req, res) => {
+      switch (process.env.ACCOUNT_PROVIDER) {
+        case 'plaid':
+          return require('../lib/plaid')
+            .createPublicToken(process.env[`PLAID_TOKEN_${nickname}`], req.body.accountNickname)
+            .then(token => res.json({ data: token }))
+            .catch(error => res.json(error))
+        default:
+          return res.json({ data: {} })
+      }
+    })
+
+    server.get('/google-sheets-url', (req, res) => {
+      return require('../lib/google/sheets')
+        .getAuthURL()
+        .then(url => res.json({ data: url }))
+        .catch(error => res.json(error))
+    })
+
+    server.get('/google-sheets-oauth2callback', (req, res) => {
+      return require('../lib/google/sheets')
+        .getToken(req.query.code)
+        .then(token => res.redirect('http://localhost:3000/sheets'))
+        .catch(error => res.json(error))
     })
 
     server.get('/', (req, res) => {
